Adding nutrition info by the selected date, and account page

+++ lib/data/model/user/user_with_profile_model.dart
@@ -6,8 +6,8 @@ class UserWithProfile {
   String lastname;
   String gender;
   int age;
-  int height;
-  int weight;
+  double height;
+  double weight;
   String activityName;
   String targetGoal;
   double bmr;
